Step3 [의존성 리팩터링] - 코드리뷰 요청드립니다. (#403)
* refactor: 피드백 반영

- Response 객체에 ofList() 추가
- 사용하지 않는 import 제거
- 조회용 Service 로직에 @Transactional 추가

* refactor: 피드백 반영

- Product 도메인에 가격 Price 불변객체로 변경

* refactor: 피드백 반영

- Menu 도메인에 가격 Price 불변객체로 변경

* refactor: 피드백 반영

- MenuProducts 생성

* refactor: 피드백 반영

- 일급컬렉션들 생성
- *Service 로직에 있는 검증들 도메인레이어로 이동

* refactor: 의존성 정리

- 도메인서비스 역할인 Validator 생성
- 일급컬렉션, 불변객체 생성
- 기존의 Service 검증로직들 Validator로 이동
- 객체들의 직접연관관계를 간접연관관계로 변경(repo를 통한 id조회)

* docs: 3단계 요구사항 정리

+++ src/main/java/kitchenpos/menu/dto/MenuProductRequest.java
@@ -17,7 +17,7 @@ public class MenuProductRequest {
 		return productId;
 	}
 
-	public int getQuantity() {
-		return quantity;
+	public Long getQuantity() {
+		return (long)quantity;
 	}
 }
